Refactor cache handler to use ES modules and move Redis client initialization inside callback
